[jOOQ/jOOQ#13053] java.lang.NoClassDefFoundError: org/postgresql/util/PGInterval when rendering SQL without the pgjdbc dependency

+++ jOOQ/src/main/java/org/jooq/impl/SelectQueryImpl.java
@@ -348,6 +348,8 @@ final class SelectQueryImpl<R extends Record> extends AbstractResultQuery<R> imp
 
 
 
+
+
     final WithImpl                                       with;
     private final SelectFieldList<SelectFieldOrAsterisk> select;
     private Table<?>                                     intoTable;
@@ -3044,9 +3046,11 @@ final class SelectQueryImpl<R extends Record> extends AbstractResultQuery<R> imp
             if (actualLimit.isApplicable()) {
                 ctx.visit(actualLimit);
             }
+
+            // [#13509] Force a LIMIT clause to prevent optimisation of "unnecessary" ORDER BY
             else if (!actualOrderBy.isEmpty() && TRUE.equals(ctx.data(DATA_FORCE_LIMIT_WITH_ORDER_BY))) {
                 Limit l = new Limit();
-                l.setLimit(Integer.MAX_VALUE);
+                l.setLimit(Long.MAX_VALUE);
                 ctx.visit(l);
             }
         }
@@ -3066,7 +3070,6 @@ final class SelectQueryImpl<R extends Record> extends AbstractResultQuery<R> imp
 
         ;
     }
-
